+++ src/data/services/questionsService.ts
@@ -258,6 +258,7 @@ export class QuestionsService extends BaseService {
    * @param {number} [params.preAnswerId] Copy question answers from this answer record.
    * @param {boolean} [params.sendToUser] If true, a notification will be sent to the user with a link to the survey.
    * @param {number} [params.notificationCategory] Send the email to organization notification user with this category.
+   * @param {number} [params.botNotificationId] Send the email to organization notification users selected by this notification ID.
    * @returns {Promise<StartSurveyAnsweringApiResponse>}
    */
   startSurveyAnswering(
@@ -269,6 +270,7 @@ export class QuestionsService extends BaseService {
       preAnswerId?: number,
       sendToUser?: boolean,
       notificationCategory?: number,
+      botNotificationId?: number,
     }
   ): Promise<StartSurveyAnsweringApiResponse> {
     if (params.locationId < 1 || isNaN(params.locationId)) {