+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning].
 ## [Unreleased]
 -
 
+## [3.3.0] - 2026-07-16
+
+### Added
+- Support zone relationship for devices
+- Add bot notifications and notification groups
+
+### Deprecated
+- Remove deprecated notification users functionality
+
 ## [3.2.1] - 2026-06-11
 
 ### Changed
@@ -84,7 +93,8 @@ and this project adheres to [Semantic Versioning].
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/peoplepower/webcore/compare/3.2.1...HEAD
+[unreleased]: https://github.com/peoplepower/webcore/compare/3.2.2...HEAD
+[3.2.2]: https://github.com/peoplepower/webcore/compare/3.2.1..3.2.2
 [3.2.1]: https://github.com/peoplepower/webcore/compare/3.2.0..3.2.1
 [3.2.0]: https://github.com/peoplepower/webcore/compare/3.1.8..3.2.0
 [3.1.8]: https://github.com/peoplepower/webcore/compare/3.1.7..3.1.8